apps/backend/pyproject.toml
[project]
name = "cyberscan-api"
version = "0.1.0"
description = "Cyberscan FastAPI backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "alembic>=1.14",
  "redis>=5.2",
  "celery>=5.4",
  "passlib[bcrypt]>=1.7.4",
  "pyjwt>=2.10",
  "python-multipart>=0.0.18",
  "httpx>=0.28",
  "dnspython>=2.7",
  "boto3>=1.35",
  "websockets>=14",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.24",
  "ruff>=0.7",
  "mypy>=1.13",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
